Update Quick Plot
I've included a file "field_values_.mat" which contains the magnitude and direction (in cylindrical coordinates) of the magnetic field.
This way you don't have to run the simulation yourself to see the results. If you want to make a quick plot of the data then load the file
in MatLab and run these three lines of code.